requirements.txt
# Copyright (C) 2020, 2023 Mitsubishi Electric Research Laboratories (MERL)
#
# SPDX-License-Identifier: AGPL-3.0-or-later

attrs==19.3.0
audioread==2.1.8
cachetools==3.1.1
catalogue==1.0.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cloudpickle==1.4.1
cycler==0.10.0
Cython==0.29.14
cytoolz==0.10.1
dask==2.17.2
future==0.18.2
gensim==3.8.0
imageio==2.8.0
imageio-ffmpeg==0.3.0
importlib-metadata==1.6.0
jmespath==0.9.4
joblib==0.15.1
jsonschema==3.2.0
kiwisolver==1.2.0
librosa==0.7.1
llvmlite==0.32.1
matplotlib==3.1.3
networkx==2.4
nltk==3.4.5
numba==0.49.1
numpy==1.18.1
pandas==1.0.1
Pillow==7.1.2
progressbar2==3.37.1
protobuf==3.11.4
python-dateutil==2.8.1
python-utils==2.3.0
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.23.0
resampy==0.2.2
rsa==4.0
s3transfer==0.3.3
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.1.0
six==1.14.0
smart-open==2.0.0
srsly==1.0.2
tensorboardX==2.0
thinc==7.3.0
toolz==0.10.0
torch==1.5.0
torchaudio==0.4.0a0+719bcc7
torchvision==0.5.0
tornado==6.0.4
tqdm==4.42.1
urllib3==1.25.8
wcwidth==0.1.9
zipp==3.1.0
pre-commit==3.2.0
